pyproject.toml
[project]
name = "audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.5.0",
    "matplotlib>=3.10.7",
    "moviepy<2",
    "numpy>=2.3.5",
    "pillow>=11.3.0",
]
